Added rough search function
Its a bit of a hit & miss since I was assuming things about a bit of what I
was working with. So it may or may not work, but it currently searches by
filename, and has the ability to be expanded to search by other things. I
don't think its being called by anything as of yet, but its there for use.