Altered error template & removed unneeded variables
the error variables in 90% of occurances were not being used, so I removed
them and moved the error message into where the error variable was being
called originally.
Because the same "we have an error" message was being displayed, along with
the <h3> and <p> formatting, I moved them to the template.