Finished refactoring the business logic to separate functions
msnak/accounts.py: Added a module for user accounts with a function to get the logged in user number, right now this hard-coded to a default value
msnak/exception.py: To communicate errors from our implementation functions back to the view functions that call them, I've made us a MediasnakError exception. Unfortunately this is used a lot.