FINALLY fixed issue #4
Fixed an issue where using a mutable as a default value in the
template_vars method was acting like a static variable. Due to its
nature, this would appear to be a caching issue, but was actually an
issue with the code itself.

This is because the default value was effectively being altered by each
call to template_vars, but due to being a default value, the bug only
manifested when the function was not passed a parameter (i.e. in the
template_with_login view).